todo_spe/todo_02/requirements.txt
requests
dateparser
typer
python-dateutil
